added color legend
Unfortunately, the GViewer is sized to the panel now, so on most
displays it will only look good when the output panel is arranged
vertically in LINQPad.

I still think the inclusion of a legend showing what the colors
represent makes this overall an improvement.

I might try to improve this further to restore the old good
automatic sizing I got when dumping the GViewer directly.